Change the approach regarding the background in the form
+ The main container will be in charge of containing the
background-color while the pseudoclass derivated from it will contain
the images. At the same type the width of the content will not be
controlled by none of those components. All of this was made to reduce
the complexity and make the styling easier. It was the easier way to fix
the never ending horizontal scrolling issue