Previous ruless for SCSS files
+ The new rules are forcing me to use old and obsolete SCSS code. @use
and @forward were new things introduced in SASS because the default
@import used in vanilla SCSS was causing issues, as result the people in
charge of the development of SASS created those alternatives. One has a
similar functionality than import while the other is a link to address
to the file. Similar to pointers in C++.
+ @mixin? Seriously, what's the point of using SASS without mixins? SASS
is something more than just create variable and wrap css rules.

+ @include? yeah, I mean considering that @mixin are givin issue, I'm
having the feeling that the are very outdated and just see SASS as a
tool to wrap selectors.

NOTE: I won't work with outdate SASS. If that is an issue with
Microverse I will take my apportunity of still being in week 04 and drop
the program. As a developer I'm pursing for being update with the tools,
I cannot show a employer a portfolio with SASS ignoring mixins and
includes, while at same time ignoring
using imports instead of the right use of @forward and @use
I DONT WANT to make them thik I see .scss files as an old .sass files with
brackets instead of indentation.